+++ so_long/libft/ft_strlen.c
@@ -6,13 +6,13 @@
 /*   By: jang-cho <jang-cho@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/07/06 18:20:25 by jang-cho          #+#    #+#             */
-/*   Updated: 2022/07/12 15:35:16 by jang-cho         ###   ########.fr       */
+/*   Updated: 2022/10/07 20:26:17 by jang-cho         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t	ft_strlen(const char *str)
+int	ft_strlen(const char *str)
 {
 	size_t	i;
